Remove configurable ID generator
